+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/In.hs
@@ -1,0 +1,27 @@
+module Cardano.Ledger.Api.Tx.In (
+  -- * Transaction input
+  TxIn (..),
+
+  -- ** Transaction ID
+  TxId (..),
+  mkTxInPartial,
+
+  -- ** Transaction index
+  TxIx,
+  txIxToInt,
+  txIxFromIntegral,
+  mkTxIxPartial,
+)
+where
+
+import Cardano.Ledger.BaseTypes (
+  TxIx,
+  mkTxIxPartial,
+  txIxFromIntegral,
+  txIxToInt,
+ )
+import Cardano.Ledger.TxIn (
+  TxId (..),
+  TxIn (TxIn),
+  mkTxInPartial,
+ )