+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `Cardano.Ledger.Api.Governance` that exposes all of the governance related types.
 * Addition of `Cardano.Ledger.Api.PParams`
 * Addition of `Cardano.Ledger.Api.Tx.Address`
+* Addition of `Cardano.Ledger.Api.Tx.In`
 * Hide `ShelleyTxBody`, `AlonzoTxBody`, `BabbageTxBody` from `Cardano.Ledger.Api.Tx.Body`
 * Export `Withdrawals` from `Cardano.Ledger.Api.Tx.Body`
 * Export `ShelleyTxAuxData`, `AllegraTxAuxData` and `AlonzoTxAuxData` from